+++ src/gray.rs
@@ -1,4 +1,4 @@
-// gray.rs      Grayscale pixel format.
+// gray.rs      Grayscale color model.
 //
 // Copyright (c) 2018-2020  Douglas P Lau
 // Copyright (c) 2019-2020  Jeron Aldaron Lau
@@ -7,18 +7,20 @@ use crate::alpha::{
     self, AChannel, Mode as _, Opaque, Premultiplied, Straight, Translucent,
 };
 use crate::gamma::{self, Linear, Srgb};
-use crate::{Ch16, Ch32, Ch8, Channel, ColorModel, Format};
+use crate::{Ch16, Ch32, Ch8, Channel, ColorModel, Pixel};
 use std::marker::PhantomData;
 use std::ops::Mul;
 
 /// Gray [color model], with optional [alpha channel].
 ///
-/// The `Channel` is either *luma* or *luminance* (depending on [gamma mode]),
-/// ranging from *black* to *white*.
+/// The `Channel` ranges from *black* to *white*.
+/// With [sRGB] gamma it is *luma*, but with [linear] gamma it is *relative
+/// luminance*.
 ///
 /// [alpha channel]: alpha/trait.AChannel.html
 /// [color model]: trait.ColorModel.html
-/// [gamma mode]: gamma/trait.Mode.html
+/// [linear]: gamma/struct.Linear.html
+/// [sRGB]: gamma/struct.Srgb.html
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 #[repr(C)]
 pub struct Gray<C, A, M, G>
@@ -28,7 +30,7 @@ where
     M: alpha::Mode,
     G: gamma::Mode,
 {
-    value: C,
+    components: [C; 1],
     alpha: A,
     mode: PhantomData<M>,
     gamma: PhantomData<G>,
@@ -47,37 +49,37 @@ where
         C: From<H>,
         A: From<Opaque<C>>,
     {
-        let value = C::from(value);
+        let components = [C::from(value)];
         let alpha = A::from(Opaque::default());
         let mode = PhantomData;
         let gamma = PhantomData;
         Gray {
-            value,
+            components,
             alpha,
             mode,
             gamma,
         }
     }
     /// Create a [Translucent](alpha/struct.Translucent.html) gray value.
-    pub fn with_alpha<H, B>(value: H, alpha: B) -> Self
+    pub fn with_alpha<H>(value: H, alpha: H) -> Self
     where
         C: From<H>,
-        A: From<B>,
+        A: From<H>,
     {
-        let value = C::from(value);
+        let components = [C::from(value)];
         let alpha = A::from(alpha);
         let mode = PhantomData;
         let gamma = PhantomData;
         Gray {
-            value,
+            components,
             alpha,
             mode,
             gamma,
         }
     }
-    /// Get the *luma* / *luminance* value.
+    /// Get the *luma* / *relative luminance* component.
     pub fn value(self) -> C {
-        self.value
+        self.components[0]
     }
     /// Get the *alpha* value.
     pub fn alpha(self) -> C {
@@ -88,12 +90,73 @@ where
 impl<C, A, M, G> ColorModel for Gray<C, A, M, G>
 where
     C: Channel,
-    A: AChannel<Chan = C>,
+    A: AChannel<Chan = C> + From<C>,
     M: alpha::Mode,
     G: gamma::Mode,
 {
-    /// `Gray` contains 2 channels (*luma* / *luminance* and *alpha*)
-    const NUM_CHANNELS: usize = 2;
+    type Chan = C;
+
+    /// Get all non-alpha components
+    fn components(&self) -> &[Self::Chan] {
+        &self.components
+    }
+
+    /// Get the *alpha* component.
+    fn alpha(self) -> Self::Chan {
+        self.alpha.value()
+    }
+
+    /// Convert to *red*, *green*, *blue* and *alpha* components
+    fn to_rgba(self) -> [Self::Chan; 4] {
+        let value = self.value();
+        [value, value, value, self.alpha()]
+    }
+
+    /// Convert from *red*, *green*, *blue* and *alpha* components
+    fn with_rgba(rgba: [Self::Chan; 4]) -> Self {
+/*        const RED_COEF: f32 = 0.2126;
+        const GREEN_COEF: f32 = 0.7152;
+        const BLUE_COEF: f32 = 0.0722;
+
+        let red = rgba[0] * RED_COEF.into();
+        let green = rgba[1] * GREEN_COEF.into();
+        let blue = rgba[2] * BLUE_COEF.into();
+        let value = red + green + blue;*/
+        let value = rgba[0].max(rgba[1]).max(rgba[2]); // FIXME
+        let alpha = rgba[3];
+        Gray::with_alpha(value, alpha)
+    }
+
+    /// Get channel-wise difference
+    fn difference(self, rhs: Self) -> Self {
+        let v = if self.value() > rhs.value() {
+            self.value() - rhs.value()
+        } else {
+            rhs.value() - self.value()
+        };
+        let a = if self.alpha() > rhs.alpha() {
+            self.alpha() - rhs.alpha()
+        } else {
+            rhs.alpha() - self.alpha()
+        };
+        Gray::with_alpha(v, a)
+    }
+
+    /// Check if all `Channel`s are within threshold
+    fn within_threshold(self, rhs: Self) -> bool {
+        self.value() <= rhs.value() && self.alpha() <= rhs.alpha()
+    }
+}
+
+impl<C, A, M, G> Pixel for Gray<C, A, M, G>
+where
+    C: Channel,
+    A: AChannel<Chan = C> + From<C>,
+    M: alpha::Mode,
+    G: gamma::Mode,
+{
+    type Alpha = M;
+    type Gamma = G;
 }
 
 impl<C, A, M, G> Iterator for Gray<C, A, M, G>
@@ -135,24 +198,24 @@ where
 impl<C, A, G> From<Gray<C, A, Straight, G>> for Gray<C, A, Premultiplied, G>
 where
     C: Channel,
-    A: AChannel<Chan = C>,
+    A: AChannel<Chan = C> + From<C>,
     G: gamma::Mode,
 {
     fn from(c: Gray<C, A, Straight, G>) -> Self {
-        let value = Premultiplied::encode(c.value, c.alpha());
-        Gray::with_alpha(value, c.alpha)
+        let value = Premultiplied::encode(c.value(), c.alpha());
+        Gray::with_alpha(value, c.alpha())
     }
 }
 
 impl<C, A, G> From<Gray<C, A, Premultiplied, G>> for Gray<C, A, Straight, G>
 where
     C: Channel,
-    A: AChannel<Chan = C>,
+    A: AChannel<Chan = C> + From<C>,
     G: gamma::Mode,
 {
     fn from(c: Gray<C, A, Premultiplied, G>) -> Self {
-        let value = Premultiplied::decode(c.value, c.alpha());
-        Gray::with_alpha(value, c.alpha)
+        let value = Premultiplied::decode(c.value(), c.alpha());
+        Gray::with_alpha(value, c.alpha())
     }
 }
 
@@ -178,140 +241,86 @@ where
 {
     type Output = Self;
     fn mul(mut self, rhs: Self) -> Self::Output {
-        self.value = self.value * rhs.value;
+        self.components[0] = self.value() * rhs.value();
         self.alpha = self.alpha * rhs.alpha;
         self
     }
 }
 
-impl<C, A, M, G> Format for Gray<C, A, M, G>
-where
-    C: Channel,
-    A: AChannel<Chan = C> + From<C>,
-    M: alpha::Mode,
-    G: gamma::Mode,
-{
-    type Chan = C;
-    type Alpha = M;
-    type Gamma = G;
-
-    /// Get *red*, *green*, *blue* and *alpha* `Channel`s
-    fn rgba(self) -> [Self::Chan; 4] {
-        [self.value, self.value, self.value, self.alpha()]
-    }
-
-    /// Make a pixel with given RGBA `Channel`s
-    fn with_rgba(rgba: [Self::Chan; 4]) -> Self {
-        let value = rgba[0].max(rgba[1]).max(rgba[2]); // FIXME
-        let alpha = rgba[3];
-        Gray::with_alpha(value, alpha)
-    }
-
-    /// Get channel-wise difference
-    fn difference(self, rhs: Self) -> Self {
-        let v = if self.value > rhs.value {
-            self.value - rhs.value
-        } else {
-            rhs.value - self.value
-        };
-        let a = if self.alpha() > rhs.alpha() {
-            self.alpha() - rhs.alpha()
-        } else {
-            rhs.alpha() - self.alpha()
-        };
-        Gray::with_alpha(v, a)
-    }
-
-    /// Check if all `Channel`s are within threshold
-    fn within_threshold(self, rhs: Self) -> bool {
-        self.value <= rhs.value && self.alpha() <= rhs.alpha()
-    }
-}
-
-/// [Gray](struct.Gray.html) 8-bit [opaque](alpha/struct.Opaque.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 8-bit [opaque](alpha/struct.Opaque.html) (no alpha)
+/// [linear](gamma/struct.Linear.html) gamma
+/// [pixel](trait.Pixel.html) format.
 pub type Gray8 = Gray<Ch8, Opaque<Ch8>, Straight, Linear>;
-/// [Gray](struct.Gray.html) 16-bit [opaque](alpha/struct.Opaque.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 16-bit [opaque](alpha/struct.Opaque.html) (no
+/// alpha) [linear](gamma/struct.Linear.html) gamma
+/// [pixel](trait.Pixel.html) format.
 pub type Gray16 = Gray<Ch16, Opaque<Ch16>, Straight, Linear>;
-/// [Gray](struct.Gray.html) 32-bit [opaque](alpha/struct.Opaque.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 32-bit [opaque](alpha/struct.Opaque.html) (no
+/// alpha) [linear](gamma/struct.Linear.html) gamma
+/// [pixel](trait.Pixel.html) format.
 pub type Gray32 = Gray<Ch32, Opaque<Ch32>, Straight, Linear>;
 
 type GrayAlpha<C, A> = Gray<C, A, Straight, Linear>;
-/// [Gray](struct.Gray.html) 8-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 8-bit [straight](alpha/struct.Straight.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha8 = GrayAlpha<Ch8, Translucent<Ch8>>;
-/// [Gray](struct.Gray.html) 16-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 16-bit [straight](alpha/struct.Straight.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha16 = GrayAlpha<Ch16, Translucent<Ch16>>;
-/// [Gray](struct.Gray.html) 32-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 32-bit [straight](alpha/struct.Straight.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha32 = GrayAlpha<Ch32, Translucent<Ch32>>;
 
 type GrayAlphap<C, A> = Gray<C, A, Premultiplied, Linear>;
 /// [Gray](struct.Gray.html) 8-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha8p = GrayAlphap<Ch8, Translucent<Ch8>>;
 /// [Gray](struct.Gray.html) 16-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha16p = GrayAlphap<Ch16, Translucent<Ch16>>;
 /// [Gray](struct.Gray.html) 32-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [linear gamma](gamma/struct.Linear.html)
-/// pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [linear](gamma/struct.Linear.html) gamma [pixel](trait.Pixel.html) format.
 pub type GrayAlpha32p = GrayAlphap<Ch32, Translucent<Ch32>>;
 
 type SGray<C, A> = Gray<C, A, Straight, Srgb>;
-/// [Gray](struct.Gray.html) 8-bit [opaque](alpha/struct.Opaque.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 8-bit [opaque](alpha/struct.Opaque.html) (no alpha)
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGray8 = SGray<Ch8, Opaque<Ch8>>;
-/// [Gray](struct.Gray.html) 16-bit [opaque](alpha/struct.Opaque.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 16-bit [opaque](alpha/struct.Opaque.html) (no
+/// alpha) [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html)
+/// format.
 pub type SGray16 = SGray<Ch16, Opaque<Ch16>>;
-/// [Gray](struct.Gray.html) 32-bit [opaque](alpha/struct.Opaque.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 32-bit [opaque](alpha/struct.Opaque.html) (no
+/// alpha) [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html)
+/// format.
 pub type SGray32 = SGray<Ch32, Opaque<Ch32>>;
 
 type SGrayAlpha<C, A> = Gray<C, A, Straight, Srgb>;
-/// [Gray](struct.Gray.html) 8-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 8-bit [straight](alpha/struct.Straight.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha8 = SGrayAlpha<Ch8, Translucent<Ch8>>;
-/// [Gray](struct.Gray.html) 16-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 16-bit [straight](alpha/struct.Straight.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha16 = SGrayAlpha<Ch16, Translucent<Ch16>>;
-/// [Gray](struct.Gray.html) 32-bit
-/// [straight alpha](alpha/struct.Straight.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [Gray](struct.Gray.html) 32-bit [straight](alpha/struct.Straight.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha32 = SGrayAlpha<Ch32, Translucent<Ch32>>;
 
 type SGrayAlphap<C, A> = Gray<C, A, Premultiplied, Srgb>;
 /// [Gray](struct.Gray.html) 8-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha8p = SGrayAlphap<Ch8, Translucent<Ch8>>;
 /// [Gray](struct.Gray.html) 16-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha16p = SGrayAlphap<Ch16, Translucent<Ch16>>;
 /// [Gray](struct.Gray.html) 32-bit
-/// [premultiplied alpha](alpha/struct.Premultiplied.html)
-/// [sRGB gamma](gamma/struct.Srgb.html) pixel [format](trait.Format.html).
+/// [premultiplied](alpha/struct.Premultiplied.html) alpha
+/// [sRGB](gamma/struct.Srgb.html) gamma [pixel](trait.Pixel.html) format.
 pub type SGrayAlpha32p = SGrayAlphap<Ch32, Translucent<Ch32>>;
 
 #[cfg(test)]