+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Renamed PremultipliedAlpha to alpha::Premultiplied
 * Mask can only have Translucent alpha (not Opaque)
 * Renamed alpha::Alpha to alpha::AChannel
+* Renamed Format trait to Pixel
 
 ## [0.9.0] - 2020-03-08
 ### Changed